+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to Bugster CLI will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.26]
+
+### Added
+- Added  `bugster auth` command for setting up or cleaning existing api key
+- Added  `-- parallel` for  `bugster run` command. Accept both  `--parallel` and  `--max-concurrent`
+
+### Changed
+- Changed max test limit to 10 
+
 ## [0.3.25]
 
 ### Changed